requirements.txt
amqp==5.3.1
asgiref==3.11.1
async-generator==1.10
attrs==25.4.0
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==6.0.3
django-filter==25.2
djangorestframework==3.17.0
drf-spectacular==0.29.0
h11==0.16.0
idna==3.11
importlib_metadata==8.7.1
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
Markdown==3.10.2
mypy_extensions==1.1.0
outcome==1.3.0.post0
packaging==26.0
prompt_toolkit==3.0.52
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
redis==7.3.0
referencing==0.37.0
rpds-py==0.30.0
selenium==4.40.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.5
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
websocket-client==1.9.0
wsproto==1.3.2
zipp==3.23.0
